// frontend-admin/src/api/generated/model/tagesabschlussStatisticsResponse.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */

export interface TagesabschlussStatisticsResponse {
  averageDailyAmount: number;
  /** @nullable */
  lastClosingDate?: string | null;
  totalAmount: number;
  totalClosings: number;
  totalTaxAmount: number;
  totalTransactions: number;
}